Add Main for Excel2Fhir
Now the process can be started via the Excel2Fhir.class with parameters for an input Excel file or directory, a temp directory for the log file and the temporarly CSV files and an output directory for the resultig JSON file(s). Additionally an option enables the splitting of the result JSON file in 1 JSON file for every patient.