Remove col "Versorgungsfall (Aufnahmegrund)"
The column "Versorgungsfall (Aufnahmegrund)" in the table sheet "Versorgungsfall" in the convertable Excel files is now removed. All diagnoses entered there in the test data Excel files were transferred to the "Diagnosen" spreadsheet and explicitly marked as admission diagnoses. This also means that the admission diagnoses in Encounter always have a correct reference to the condition with the corresponding diagnosis code and date.