Add one Diagnosis to every Sub-Encounter
In order to create valid packages now one of all diagnosis will be added as DiagnosisComponent to all Part-Of-Encounters. Which one is determined by the diagnosis role. The Chief Complaint (CC) is prefererred. If not exsts then the first found Comorbidity diagnosis (CM) will be choosen. If not exists... -> see comment in Diagnosis_Role.map.