Prevent Invalid References To Medication
If a Medication Resource was invalid, then a MedicationStatement or a MedicationAdministration Resource was always built anyway, which contained a reference to the Medication Resource. However, since the invalid Medication Resource is not included in the bundle, the objects pointing to it are not allowed to be included in the bundle either.