Resorces are invalid if Patient or Main Encounter Reference is broken
Now only resources are in the result bundle which have a valid reference to the patient and/or the main encounter. If the patient or main encounter is invalid (and so not in the bundle), then now no Medication/Procedure/etc... will have reference to the not existing patient or enounter.