Correct Encounters Class Code
The class code will be read now from a resource file and is no longer hard coded. Both encounters have now the same class code that is given to the main encounter. So the partOf encounter gets this class code in the bundle post processor equivalent to the missing diaognosis.