Change File Name Creation
If there are lesser patients in the dataset than should be written to a single file, then the patient IDs will no longer be appended to the file name. Only the file name prefix specified in the conversion options in the Excel file is appended along with the suffix. This ensures stable file names when patients are added.