Add both location types to Encounter Level 3 (real and "fake" ref)
Encounter can have a location reference to a real Location resource object.
The alternative way is that the Reference resource object does not point to an existing Location resource but has an Identifier and a display only (means in my words "fake" reference) and the physicalType of the encounter location concepts must be set.
Now both ways are implemented: There is a real Reference object with a valid reference to an existing Location object that has its own physicalType and additionally the Reference has an Identifier and a display and the physicalType is also set for the Encounter location concept.